requirements.txt
aiogram==3.13.1
aiohttp==3.10.10
python-dotenv==1.0.1
curl_cffi==0.7.3
phonenumbers==8.13.39
asyncwhois==1.1.12
aiosqlite==0.20.0
dnspython==2.6.1
